Moved conversion SDK->Jenkins models conversion methods to Utils class from UtilsCompat
