+++ src/main/java/com/dynatrace/jenkins/dashboard/utils/UtilsCompat.java
@@ -44,66 +44,6 @@ public final class UtilsCompat {
 	private UtilsCompat() {
 	}
 
-	public static TAReportDetails convertTestRuns(TestRuns sdkTestRuns) {
-		ArrayList<TestRun> testRuns = new ArrayList<>();
-		if (sdkTestRuns != null) {
-			for (com.dynatrace.sdk.server.testautomation.models.TestRun tr : sdkTestRuns.getTestRuns()) {
-				testRuns.add(convertTestRun(tr));
-			}
-		}
-		return new TAReportDetails(testRuns);
-	}
-
-	public static TestRun convertTestRun(com.dynatrace.sdk.server.testautomation.models.TestRun sdkTestRun) {
-		List<TestResult> testResults = new ArrayList<>();
-		for (com.dynatrace.sdk.server.testautomation.models.TestResult sdkResult : sdkTestRun.getTestResults()) {
-			testResults.add(convertTestResult(sdkResult));
-		}
-		Map<TestStatus, Integer> testRunSummary = new EnumMap<>(TestStatus.class);
-		testRunSummary.put(TestStatus.FAILED, sdkTestRun.getFailedCount());
-		testRunSummary.put(TestStatus.DEGRADED, sdkTestRun.getDegradedCount());
-		testRunSummary.put(TestStatus.VOLATILE, sdkTestRun.getVolatileCount());
-		testRunSummary.put(TestStatus.IMPROVED, sdkTestRun.getImprovedCount());
-		testRunSummary.put(TestStatus.PASSED, sdkTestRun.getPassedCount());
-		return new TestRun(testResults, testRunSummary, sdkTestRun.getId(), convertTestCategory(sdkTestRun.getCategory()));
-	}
-
-	public static TestResult convertTestResult(com.dynatrace.sdk.server.testautomation.models.TestResult sdkTestResult) {
-		Set<TestMeasure> measures = new HashSet<>();
-		for (com.dynatrace.sdk.server.testautomation.models.TestMeasure sdkMeasure : sdkTestResult.getMeasures()) {
-			measures.add(convertTestMeasure(sdkMeasure));
-		}
-		return new TestResult(new Date(sdkTestResult.getExecutionTime()), sdkTestResult.getName(), sdkTestResult.getPackageName(), sdkTestResult.getPlatform(), convertTestStatus(sdkTestResult.getStatus()), measures);
-	}
-
-	public static TestMeasure convertTestMeasure(com.dynatrace.sdk.server.testautomation.models.TestMeasure sdkTestMeasure) {
-		return new TestMeasure(sdkTestMeasure.getName(),
-				sdkTestMeasure.getMetricGroup(),
-				sdkTestMeasure.getExpectedMin(),
-				sdkTestMeasure.getExpectedMax(),
-				sdkTestMeasure.getValue(),
-				sdkTestMeasure.getUnit(),
-				sdkTestMeasure.getViolationPercentage());
-	}
-
-	public static TestCategory convertTestCategory(com.dynatrace.sdk.server.testautomation.models.TestCategory sdkTestCategory) {
-		switch (sdkTestCategory) {
-			case UNIT:
-				return TestCategory.UNIT;
-			case UI_DRIVEN:
-				return TestCategory.UI_DRIVEN;
-			case WEB_API:
-				return TestCategory.WEB_API;
-			case PERFORMANCE:
-				return TestCategory.PERFORMANCE;
-		}
-		throw new IllegalArgumentException("Could not convert TestCategory");
-	}
-
-	public static TestStatus convertTestStatus(com.dynatrace.sdk.server.testautomation.models.TestStatus sdkTestStatus) {
-		return TestStatus.valueOf(sdkTestStatus.name());
-	}
-
 	public static TAReport getCompatReport(AbstractBuild<?, ?> build) {
 		final TestAutomationBuildAction oldBuildAction = build.getAction(TestAutomationBuildAction.class);
 		if (oldBuildAction == null) {
